Merge

+++ src/main/java/com/qbit/exchanger/money/model/Transfer.java
@@ -53,11 +53,11 @@ public class Transfer implements Serializable {
 		this.amount = amount;
 	}
 
-	public int getCoins() {
+	public long getCoins() {
 		return amount.getCoins();
 	}
 
-	public int getCents() {
+	public long getCents() {
 		return amount.getCents();
 	}
